LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/js/CSS-supports-selector-picker-expected.txt


PASS ::picker-icon pseudo-element is supported
PASS ::picker(select) pseudo-element is supported
PASS ::picker-icon without element is supported
PASS ::picker(select) without element is supported
PASS ::picker without argument is not supported
PASS ::picker with empty argument is not supported
